advanced-examples/news-scout/requirements.txt
# News Scout Dependencies

# Core dependencies
PyYAML>=6.0.1
requests>=2.31.0

# RSS parsing
feedparser>=6.0.10

# Text processing
beautifulsoup4>=4.12.0

# Optional: Advanced NLP
# nltk>=3.8.1
# spacy>=3.7.0
# transformers>=4.35.0  # For advanced summarization

# Optional: Sentiment analysis
# textblob>=0.17.1
# vaderSentiment>=3.3.2

# Database
# sqlite3 is built-in